pyproject.toml
[tool.poetry]
name = "ibooks_notion_pipeline"
version = "0.1.0"
description = "Sync iBooks annotations to Notion"
authors = ["fstuelzebach"]
readme = "README.md"

packages = [{ include = "ibooks_notion_pipeline" }]

[tool.poetry.dependencies]
python = "^3.10"
python_common = { path = "../../python_common", develop = true }
requests = "^2.0"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
